# schema_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "13.0.5"
  },
  "host": "ezemaacyyvbpjlagchds.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.orders.id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_price"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.status"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.items"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.dish_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.delivery_address"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_cents"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.order_value"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.confirmed_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/orders"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.orders"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.orders.id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_price"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.status"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.items"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.dish_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.delivery_address"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_cents"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.order_value"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.confirmed_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.orders.id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_price"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.status"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.items"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.restaurant_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.dish_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_name"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.customer_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.delivery_address"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.notes"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_cents"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.total_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.order_value"
          },
          {
            "$ref": "#/parameters/rowFilter.orders.confirmed_at"
          },
          {
            "$ref": "#/parameters/body.orders"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/restaurants": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.restaurants.id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.name"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.address"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.city"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.cuisine_type"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lat"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lng"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_email"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.partner_mode"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.menu_count"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_active"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.opening_hours"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_open"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.min_order_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.delivery_available"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/restaurants"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "restaurants"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.restaurants"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "restaurants"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.restaurants.id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.name"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.address"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.city"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.cuisine_type"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lat"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lng"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_email"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.partner_mode"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.menu_count"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_active"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.opening_hours"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_open"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.min_order_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.delivery_available"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "restaurants"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.restaurants.id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.name"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.address"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.city"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.cuisine_type"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lat"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.lng"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.owner_email"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.business_id"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.partner_mode"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.menu_count"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_active"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.opening_hours"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.is_open"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.min_order_pln"
          },
          {
            "$ref": "#/parameters/rowFilter.restaurants.delivery_available"
          },
          {
            "$ref": "#/parameters/body.restaurants"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "restaurants"
        ]
      }
    },
...
  },
...
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.orders": {
      "name": "orders",
      "description": "orders",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/orders"
      }
    },
    "rowFilter.orders.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.restaurant_id": {
      "name": "restaurant_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.user_id": {
      "name": "user_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.total_price": {
      "name": "total_price",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.items": {
      "name": "items",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.restaurant_name": {
      "name": "restaurant_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.dish_name": {
      "name": "dish_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.customer_name": {
      "name": "customer_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.customer_phone": {
      "name": "customer_phone",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.delivery_address": {
      "name": "delivery_address",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.notes": {
      "name": "notes",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.business_id": {
      "name": "business_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.total_cents": {
      "name": "total_cents",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.total_pln": {
      "name": "total_pln",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.order_value": {
      "name": "order_value",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.orders.confirmed_at": {
      "name": "confirmed_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "rowFilter.restaurants.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.name": {
      "name": "name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.address": {
      "name": "address",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.city": {
      "name": "city",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.cuisine_type": {
      "name": "cuisine_type",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.rating": {
      "name": "rating",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.image_url": {
      "name": "image_url",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.restaurants.owner_id": {
      "name": "owner_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}